requirements.txt
fastapi
uvicorn
openai
pandas
python-dotenv
tiktoken